animate/ARAP_standalone/requirements.txt
certifi==2020.12.5
cycler==0.10.0
decorator==4.4.2
freetype-py==2.2.0
glfw==2.1.0
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
multipledispatch==0.6.0
networkx==2.5.1
numpy==1.20.2
Pillow==8.2.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyrr==0.10.3
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
threadpoolctl==2.1.0
tifffile==2021.4.8
triangle==20200424
